Fix how intensity is found.
To find light intensity on a given triangle ABC, we need to find its normal
to perform a dot product with the light direction vector. The normal was
previously found by making a cross product between the vector OA and OB
(O being origin), which is wrong. It is supposed to
be the cross product between AB and AC that gives us the correct normal.